requirements.txt
opencv-python
numpy
langchain
langchain-google-genai
chromadb
python-dotenv
reportlab
pypdf
